windows

+++ sql/sql_statistics.cc
@@ -1058,7 +1058,7 @@ public:
           else
           {
             table_field->collected_stats->min_value->val_str(&val);
-            uint32 length= Well_formed_prefix(val.charset(), val.ptr(),
+            size_t length= Well_formed_prefix(val.charset(), val.ptr(),
                            MY_MIN(val.length(), stat_field->field_length)).length();
             stat_field->store(val.ptr(), length, &my_charset_bin);
           }
@@ -1069,7 +1069,7 @@ public:
           else
           {
             table_field->collected_stats->max_value->val_str(&val);
-            uint32 length= Well_formed_prefix(val.charset(), val.ptr(),
+            size_t length= Well_formed_prefix(val.charset(), val.ptr(),
                             MY_MIN(val.length(), stat_field->field_length)).length();
             stat_field->store(val.ptr(), length, &my_charset_bin);
           }
